+++ src/states/SingleSupport.h
@@ -74,15 +74,14 @@ namespace lipm_walking
       void updatePreview();
 
     private:
-      SwingFoot swingFoot_;
-      bool hasUpdatedMPCOnce_;
-      double duration_;
-      double earlyDoubleSupportDuration_;
-      double remTime_;
-      double stateTime_;
-      double timeSinceLastPreviewUpdate_;
-      std::shared_ptr<mc_tasks::force::CoPTask> supportFootTask;
-      std::shared_ptr<mc_tasks::force::CoPTask> swingFootTask;
+      SwingFoot swingFoot_; /**< Swing foot trajectory interpolator */
+      bool hasUpdatedMPCOnce_; /**< Has the walking pattern been updated since the beginning of the SSP? */
+      double duration_; /**< Total duration of the SSP in [s] */
+      double remTime_; /**< Time remainin guntil the end of the phase */
+      double stateTime_; /** Time since the beginning of the SSP */
+      double timeSinceLastPreviewUpdate_; /**< Time count used to schedule MPC udpates, in [s] */
+      std::shared_ptr<mc_tasks::force::CoPTask> supportFootTask; /**< Current support foot task from the stabilizer */
+      std::shared_ptr<mc_tasks::force::CoPTask> swingFootTask; /**< Current swing foot task from the stabilizer */
     };
   }
 }